Aded FakeModule, DashBoard and WindowRegion
Huge checkin, opps.
Removed unused images.
Decided that each main view would be an independant window. Updated
Region behavior to do this.
Created the DashboardModule which is the core module for actually shoing
Activated profiles.
Moved the Settings viw to its own window (which need plaety of work
still).
This is now effectively a Shell-less design, ie. there is no MainWindow
for the Application. It will be more a service with Windws that popup
when required.